# docs/swagger.yaml
definitions:
...
  dto.PageResponse:
    properties:
      count:
        description: Total number of items that match the query
        type: integer
      next:
        description: URL to the next page of results
        type: string
      num_pages:
        description: Total number of pages that match the query
        type: integer
      per_page:
        description: Total number of items per page that match the query
        type: integer
      previous:
        description: URL to the previous page of results
        type: string
    type: object
  dto.SaltCachePageResponse:
    properties:
      paging:
        $ref: '#/definitions/dto.PageResponse'
      results:
        description: Array of salt cache results
        items:
          $ref: '#/definitions/model.SaltCache'
        type: array
    type: object
...
  github_com_jackc_pgx_pgtype.JSONB:
    properties:
      bytes:
        items:
          type: integer
        type: array
      status:
        $ref: '#/definitions/github_com_jackc_pgx_pgtype.Status'
    type: object
  github_com_jackc_pgx_pgtype.Status:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - Undefined
    - "Null"
    - Present
  httputil.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
...
  model.SaltCache:
    properties:
      bank:
        type: string
      data:
        $ref: '#/definitions/github_com_jackc_pgx_pgtype.JSONB'
      dataChanged:
        type: string
      id:
        type: string
      psqlkey:
        type: string
    type: object
...
info:
  contact:
    email: pmartin@gatech.edu
    name: API Support
    url: http://oit.gatech.edu
  description: This the Agartha API Backend
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Agartha API
  version: "1.0"
paths:
...
  /api/v1/salt_cache:
    get:
      consumes:
      - application/json
      description: Retrieve a list of salt_cache items filtered by bank and key with
        pagination support.
      parameters:
      - description: Bank of the cached item to search (Supports wildcards * and ?
          for single char matches.)
        in: query
        name: bank
        type: string
      - description: Key of the cached item to search (Supports wildcards * and ?
          for single char matches.)
        in: query
        name: key
        type: string
      - description: Load the data field. This defaults to false for performance reasons
        in: query
        name: load_data
        type: boolean
      - description: Filter items from this date (RFC3339 format).
        in: query
        name: since
        type: string
      - description: Filter items up to this date (RFC3339 format).
        in: query
        name: until
        type: string
      - description: Page number of results to retrieve
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SaltCachePageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - Bearer: []
      tags:
      - SaltCache
  /api/v1/salt_cache/{bank}/{key}:
    delete:
      consumes:
      - application/json
      description: Delete one salt_cache item by bank and key
      parameters:
      - description: bank of the salt cache item to delete
        in: path
        name: bank
        required: true
        type: string
      - description: key of the salt cache item to delete
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SaltCache'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - Bearer: []
      tags:
      - SaltCache
    get:
      consumes:
      - application/json
      description: Get one salt_cache item by bank and key
      parameters:
      - description: bank of the salt cache item to retrieve
        in: path
        name: bank
        required: true
        type: string
      - description: key of the salt cache item to retrieve
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SaltCache'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - Bearer: []
      tags:
      - SaltCache
  /api/v1/salt_cache/uuid/{uuid}:
    delete:
      consumes:
      - application/json
      description: Delete one salt_cache item by UUID
      parameters:
      - description: uuid of the salt cache item to delete
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SaltCache'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - Bearer: []
      tags:
      - SaltCache
    get:
      consumes:
      - application/json
      description: Get one salt_cache item by UUID
      parameters:
      - description: uuid of the salt cache item to retrieve
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SaltCache'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - Bearer: []
      tags:
      - SaltCache
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"